Move structuralize pass to separate file
Separate structuralize pass logic to handle two cases:
1. Rewrite as generic version
2. Handle very simple custom versions

Other custom versions will be left unchanged and will need to be
handled by another lowering pass which only looks for those specific
custom versions.
